rename: EpisodeCostPolicy에서 CoinCostPolicy로 변경
