+++ src/main/java/com/ham/netnovel/coinCostPolicy/data/RangeFrom.java
@@ -1,4 +1,4 @@
-package com.ham.netnovel.episodeCostPolicy.data;
+package com.ham.netnovel.coinCostPolicy.data;
 
 public enum RangeFrom {
     RECENT, //최신 회차부터
